TwoPointFunctions.jl

TwoPointFunctions.cosχ — Function

cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Cosine of the angle between the unit vectors denoted by the spherical coordinates $(\theta_1,\phi_1)$ and $(\theta_2,\phi_2)$. This is equivalent to the cosine of the angular distance along a great circle between these two points on a sphere.

Numerically this is given by

\[\cos\chi((\theta_1,\phi_1), (\theta_2,\phi_2)) = \cos(\theta_1)\cos(\theta_2) + \sin(\theta_1)\sin(\theta_2)\cos(\phi_1 - \phi_2)\]

The unicode character at the end of the function's name is \chi, and not x.

Examples

julia> cosχ((0,0), (0,0))
1.0

julia> cosχ((0,0), (pi,0))
-1.0

TwoPointFunctions.∂²θ₁cosχ — Method

∂²θ₁cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Second derivative of cosχ with respect to θ₁.

Examples

julia> ∂²θ₁cosχ((0,0), (0,0))
-1.0

julia> ∂²θ₁cosχ((0,0), (pi,0))
1.0

TwoPointFunctions.∂²θ₂cosχ — Method

∂²θ₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Second derivative of cosχ with respect to θ₂.

Examples

julia> ∂²θ₂cosχ((0,0), (0,0))
-1.0

julia> ∂²θ₂cosχ((0,0), (pi,0))
1.0

TwoPointFunctions.∂²ϕ₁cosχ — Function

∂²ϕ₁cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Second derivative of cosχ with respect to ϕ₁.

Examples

julia> ∂²ϕ₁cosχ((0,0), (0,0))
-0.0

julia> ∂²ϕ₁cosχ((pi/2,0), (pi/2, 0))
-1.0

TwoPointFunctions.∂²ϕ₂cosχ — Function

∂²ϕ₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Second derivative of cosχ with respect to ϕ₂.

Examples

julia> ∂²ϕ₂cosχ((0,0), (0,0))
-0.0

julia> ∂²ϕ₂cosχ((pi/2,0), (pi/2, 0))
-1.0

TwoPointFunctions.∂θ₁cosχ — Function

∂θ₁cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Partial derivative of cosχ with respect to θ₁.

Examples

julia> ∂θ₁cosχ((0,0), (0,0))
0.0

julia> ∂θ₁cosχ((0,0), (pi/2,0))
1.0

See also: ∂θ₂cosχ

TwoPointFunctions.∂θ₁∂θ₂cosχ — Function

TwoPointFunctions.∂θ₁∂θ₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Mixed derivative of cosχ with respect to θ₁ and θ₂.

Examples

julia> TwoPointFunctions.∂θ₁∂θ₂cosχ((0,0), (0,0))
1.0

julia> TwoPointFunctions.∂θ₁∂θ₂cosχ((0,0), (pi,0))
-1.0

TwoPointFunctions.∂θ₂cosχ — Function

∂θ₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Partial derivative of cosχ with respect to θ₂.

Examples

julia> ∂θ₂cosχ((0,0), (0,0))
0.0

julia> ∂θ₂cosχ((0,0), (pi/2,pi/2))
-1.0

See also: ∂θ₁cosχ

TwoPointFunctions.∂ϕ₁cosχ — Function

∂ϕ₁cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Partial derivative of cosχ with respect to ϕ₁.

Examples

julia> ∂ϕ₁cosχ((pi/2, 0), (pi/2, pi/3)) ≈ sin(pi/3)
true

julia> ∂ϕ₁cosχ((pi/2, 0), (pi/2, pi/2))
1.0

See also: ∂ϕ₂cosχ

TwoPointFunctions.∂ϕ₂cosχ — Function

∂ϕ₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Partial derivative of cosχ with respect to ϕ₂.

Examples

julia> ∂ϕ₂cosχ((pi/2, 0), (pi/2, pi/3)) ≈ -sin(pi/3)
true

julia> ∂ϕ₂cosχ((pi/2, 0), (pi/2, pi/2))
-1.0

See also: ∂ϕ₁cosχ

TwoPointFunctions.∇ϕ₁cosχ — Method

∇ϕ₁cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

The ${\hat{\phi}}_1$ component of $\nabla_1 \cos \chi$ in the spherical polar basis. Numerically this is equivalent to

\[\nabla \phi_1 \cos\chi ((\theta_1, \phi_1), (\theta_2, \phi_2)) = \frac{1}{\sin(\theta_1)} \partial \phi_1 \cos\chi ((\theta_1, \phi_1), (\theta_2, \phi_2))\]

except ∇ϕ₁cosχ is defined to avoid coordinate singularities at the poles.

Examples

julia> ∇ϕ₁cosχ((0, 0), (0, 0))
-0.0

julia> ∇ϕ₁cosχ((pi/2, 0), (pi/2, pi/2))
1.0

See also: ∂ϕ₁cosχ, ∇ϕ₂cosχ

TwoPointFunctions.∇ϕ₂cosχ — Method

∇ϕ₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

The ${\hat{\phi}}_2$ component of $\nabla_2 \cos \chi$ in the spherical polar basis. Numerically this is equivalent to

\[\nabla \phi_2 \cos\chi ((\theta_1, \phi_1), (\theta_2, \phi_2)) = \frac{1}{\sin(\theta_2)} \partial \phi_2 \cos\chi ((\theta_1, \phi_1), (\theta_2, \phi_2))\]

except ∇ϕ₂cosχ is defined to avoid coordinate singularities at the poles.

Examples

julia> ∇ϕ₂cosχ((0, 0), (0, 0))
0.0

julia> ∇ϕ₂cosχ((pi/2, 0), (pi/2, pi/2))
-1.0

See also: ∂ϕ₂cosχ, ∇ϕ₁cosχ

TwoPointFunctions.∇₁cosχ — Function

∇₁cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Return $(\hat{\theta}_1 \cdot \nabla_1 \cos \chi,\, \hat{\phi}_1 \cdot \nabla_1 \cos \chi)$, where ${\nabla}_1$ represents the gradient on the unit sphere with respect to $(\theta_1, \phi_1)$.

Examples

julia> ∇₁cosχ((0,0), (pi/2, pi/3)) == (cos(pi/3), sin(pi/3))
true

TwoPointFunctions.∇₂cosχ — Function

∇₂cosχ((θ₁,ϕ₁), (θ₂,ϕ₂))

Return $(\hat{\theta}_2 \cdot \nabla_2 \cos \chi,\, \hat{\phi}_2 \cdot \nabla_2 \cos \chi)$, where ${\nabla}_2$ represents the gradient on the unit sphere with respect to $(\theta_2, \phi_2)$.

Examples

julia> ∇₂cosχ((0,0), (pi/2, pi/3))
(-1.0, -0.0)